requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
Django==4.0.3
django-appconf==1.0.5
django-bootstrap4==21.2
django-classy-tags==3.0.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-sslserver==0.22
django-tinymce==3.4.0
djangorestframework==3.13.1
gunicorn==20.1.0
numpy==1.22.2
Pillow==9.0.1
pytz==2022.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
